refactor: Standardize types organization and file naming
- Extract types from about/data.ts into dedicated types/Data.ts file
- Standardize file naming convention to use capital 'D' in Data.ts
- Update imports to reflect new file structure
- Maintain consistent type organization across about and resume sections
